+++ python/aristotle-metadata-registry/aristotle_mdr/contrib/slots/manager.py
@@ -1,6 +1,7 @@
 from django.db.models import Manager
 from django.db.models.query import QuerySet
-from aristotle_mdr.contrib.slots.choices import permission_choices as perms
+
+from aristotle_mdr.constants import visibility_permission_choices as permission_choices
 
 
 class SimplePermsQueryset(QuerySet):
@@ -19,13 +20,13 @@ class SimplePermsQueryset(QuerySet):
             else:
                 # Return public and auth only slots
                 kwargs = {
-                    self.perm_field_in: [perms.public, perms.auth]
+                    self.perm_field_in: [permission_choices.public, permission_choices.auth]
                 }
                 return self.filter(**kwargs)
         else:
             # Only return public slots
             kwargs = {
-                self.perm_field_name: perms.public
+                self.perm_field_name: permission_choices.public
             }
             return self.filter(**kwargs)