+++ python/aristotle-metadata-registry/aristotle_mdr/templatetags/util_tags.py
@@ -15,6 +15,11 @@ register = template.Library()
 
 
 @register.filter
+def getdir(obj):
+    return dir(obj)
+
+
+@register.filter
 def order_by(qs, order):
     return qs.order_by(*(order.split(",")))